+++ promptolution/templates.py
@@ -91,8 +91,7 @@ OPRO_TEMPLATE = """Your task is to generate an instruction.
 
 Below are some previous instructions with their scores. The score ranges from 0 to 100.
 
-<old_instructions>
-
+<instructions>
 Here are some examples of the target dataset:
 <examples>
 
@@ -105,8 +104,7 @@ OPRO_TEMPLATE_TD = """Your task is to generate an instruction for the following 
 
 Below are some previous instructions with their scores. The score ranges from 0 to 100.
 
-<old_instructions>
-
+<instructions>
 Here are some examples of the target dataset:
 <examples>